Add codec tests file and move existing codec tests into it
