+++ tests/test_multiaddr.py
@@ -2,8 +2,6 @@
 # -*- coding: utf-8 -*-
 import pytest
 
-from multiaddr.codec import string_to_bytes
-from multiaddr.codec import bytes_to_string
 from multiaddr.multiaddr import Multiaddr
 from multiaddr.multiaddr import ProtocolNotFoundException
 from multiaddr.protocols import protocol_with_name
@@ -101,21 +99,6 @@ def test_eq():
     assert m4 == m2
     assert m3 == m4
     assert m4 == m3
-
-
-def test_string_to_bytes():
-    assert string_to_bytes("/ip4/127.0.0.1/udp/1234") == b'047f0000011104d2'
-    assert string_to_bytes("/ip4/127.0.0.1/tcp/4321") == b'047f0000010610e1'
-    assert (
-        string_to_bytes("/ip4/127.0.0.1/udp/1234/ip4/127.0.0.1/tcp/4321") ==
-        b'047f0000011104d2047f0000010610e1')
-
-
-def test_bytes_to_string():
-    assert bytes_to_string(b"047f0000011104d2") == "/ip4/127.0.0.1/udp/1234"
-    assert bytes_to_string(b"047f0000010610e1") == "/ip4/127.0.0.1/tcp/4321"
-    assert (bytes_to_string(b"047f0000011104d2047f0000010610e1") ==
-            "/ip4/127.0.0.1/udp/1234/ip4/127.0.0.1/tcp/4321")
 
 
 @pytest.mark.parametrize(
